优化调整

+++ xmfcn-spring-cloud-common/src/main/java/com/cn/xmf/base/model/Partion.java
@@ -120,7 +120,7 @@ public class Partion implements java.io.Serializable {
      * @author rufei.cn
      * @return
      */
-    public long getPageCount() {
+    public int getPageCount() {
         return pageCount;
     }
